requirements.txt
openai
PyMuPDF
Pillow
pytesseract
